Changes to uboot and other low level things. Added fbutils package too
Added a patch to place a new config for uboot into the build directory
added an append to the platform-init to use custom code to intilize, and
the config above then uses this.
Added fbutils package, and configured the gpios on startup that we need.
(other functions in this package are not touched currently)