First bash at the REST api for the Track1 boards
Has modified BMC node,
modifed server to support TPDBs, with LED, EEPROM and PWM control. LEDs
and PWMs are leaf nodes.

structure is /api/mezz/(bmc)|(slotn) and it gets filled in as slots were
allocated.

None of the FRUID stuff works, as we've not defined our fruid struture.
Needs to be done.