Full working version of the xvServer and bit file.
This is a tested version of the Xilinx Virtual Cable interface

HOW-TO
======

ssh into bmc, and enable the jtag system
'jtag-setup.s --oe --ethernet --reset 1 2 3 4'

Run the server
'xvcServer'

On the PC you wish to use, start xsbc, and attach a virtual link
'connect -xvc-url tcp:10.10.1.19:2542'
where 10.10.1.19 is the ip address of the MicroZed

xsdb will tell you it has created a new target! Vivado will show this as
a new pod, you can just use it. Or you can connect to it just like a
diligent pod and use the xsdb commandline tools.

With 4 QFDBs the chain is very long! And things are running around 3ish
Mhz, so it's not as fast as a direct connection.

When you are done,
kill the xvcServer wth a ctrl-c or so

put jtag back to standby

'jtag-setup.s --nooe --conector --reset'