Serial via ssh added.
4 users added, qfdb_a/b/c/d.
These then directly access the sol-util for the module based on their
name, giving direct access to the serial links.
On serial disconnect they drop the ssh connection.

Patched the mterm_server to correctly set the sockets it uses to allow
the new users to access via the corect group.