Added in a firmware setup util
A Python util to set up the track1 eeprom (when we do shipping)

Allows us to set SerialNo., MAC, hardware revisions and also default
config for 4 QFDBs.

Any mac given will be made local administered, with the lower octet
using part of the slot ID.

Util is written in python.