Front Panel daemon added
This will monitor the health and turn on the alarm if needed.
it also allows a flashing identify mode.

Monitors the button, a quick press is a module reset, a long press is a
power operation.

When all are off, everything gets turned on.
if any are on, all are turned off.