Added REST access to the bootmode settings via a node
It exists in /api/mezz/bootmode

It will show the current stored status, QSPI, JTAG or UNKNOWN for each
site.

Setting it uses the site name qfdb_a etc and "JTAG" or "QSPI" to set it.

It stores the variable for persitsance, it also sets the hardware to
match so that a node reboot/reset will use the new value.